+++ Modules/TestQuestionPool/classes/class.assQuestionGUI.php
@@ -13,7 +13,7 @@ require_once 'Modules/Test/classes/class.ilTestExpressPage.php';
 *
 * @author		Helmut Schottmüller <helmut.schottmueller@mac.com>
 * @author		Björn Heyser <bheyser@databay.de>
-* @version		$Id: class.assQuestionGUI.php 47546 2014-01-27 12:46:25Z bheyser $
+* @version		$Id$
 * @ingroup		ModulesTestQuestionPool
 */
 abstract class assQuestionGUI
@@ -309,8 +309,7 @@ abstract class assQuestionGUI
 		$page_gui->setQuestionHTML(array($this->object->getId() => $html));
 		$page_gui->setOutputMode("presentation");
 		$presentation = $page_gui->presentation();
-		// bugfix for non XHTML conform img tags in ILIAS Learning Module Editor
-		$presentation = preg_replace("/src=\".\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
+		$presentation = preg_replace("/src=\"\\.\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
 		return $presentation;
 	}
 
@@ -340,7 +339,7 @@ abstract class assQuestionGUI
 		
 		if( $this->object->areObligationsToBeConsidered() && ilObjTest::isQuestionObligatory($this->object->getId()) )
 		{
-			$obligatoryString = ' *';
+			$obligatoryString = '([-_-])';
 		}
 		else
 		{
@@ -371,8 +370,13 @@ abstract class assQuestionGUI
 		}
 		$presentation = $page_gui->presentation();
 		if (strlen($maxpoints)) $presentation = str_replace($maxpoints, "<em>$maxpoints</em>", $presentation);
-		// bugfix for non XHTML conform img tags in ILIAS Learning Module Editor
-		$presentation = preg_replace("/src=\".\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
+		if (strlen($obligatoryString))
+		{
+			$replacement	='<br><span class="obligatory" style="font-size:small">'.
+				$this->lng->txt("tst_you_have_to_answer_this_question").'</span>';
+			$presentation 	= str_replace($obligatoryString, $replacement, $presentation);
+		}
+		$presentation = preg_replace("/src=\"\\.\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
 		return $presentation;
 	}
 	
@@ -405,7 +409,11 @@ abstract class assQuestionGUI
 		}
 	}
 
-	function originalSyncForm($return_to = "")
+	/**
+	 * @param string $return_to
+	 * @param string $return_to_feedback  ilAssQuestionFeedbackEditingGUI
+	 */
+	function originalSyncForm($return_to = "", $return_to_feedback = '')
 	{
 		if (strlen($return_to))
 		{
@@ -414,6 +422,11 @@ abstract class assQuestionGUI
 		else if ($_REQUEST['return_to']) {
 			$this->ctrl->setParameter($this, "return_to", $_REQUEST['return_to']);
 		}
+		if(strlen($return_to_feedback))
+		{
+			$this->ctrl->setParameter($this, 'return_to_fb', 'true');
+		}	
+		
 		$template = new ilTemplate("tpl.il_as_qpl_sync_original.html",TRUE, TRUE, "Modules/TestQuestionPool");
 		$template->setVariable("BUTTON_YES", $this->lng->txt("yes"));
 		$template->setVariable("BUTTON_NO", $this->lng->txt("no"));
@@ -433,6 +446,10 @@ abstract class assQuestionGUI
 		{
 			$this->ctrl->redirect($this, $_GET["return_to"]);
 		}
+		if (strlen($_REQUEST["return_to_fb"]))
+		{
+			$this->ctrl->redirectByClass('ilAssQuestionFeedbackEditingGUI', 'show');
+		}
 		else
 		{
 			$_GET["ref_id"] = $_GET["calling_test"];
@@ -445,6 +462,10 @@ abstract class assQuestionGUI
 		if (strlen($_GET["return_to"]))
 		{
 			$this->ctrl->redirect($this, $_GET["return_to"]);
+		}
+		if(strlen($_REQUEST['return_to_fb']))
+		{
+			$this->ctrl->redirectByClass('ilAssQuestionFeedbackEditingGUI', 'show');
 		}
 		else
 		{
@@ -914,7 +935,7 @@ abstract class assQuestionGUI
 		$question->setCols(80);
 		if (!$this->object->getSelfAssessmentEditingMode())
 		{
-			if( $this->object->getAdditionalContentEditingMode() == assQuestion::ADDITIONAL_CONTENT_EDITING_MODE_DEFAULT )
+			if( $this->object->getAdditionalContentEditingMode() != assQuestion::ADDITIONAL_CONTENT_EDITING_MODE_PAGE_OBJECT )
 			{
 				$question->setUseRte(TRUE);
 				include_once "./Services/AdvancedEditing/classes/class.ilObjAdvancedEditing.php";
@@ -1223,12 +1244,15 @@ abstract class assQuestionGUI
 				$file->setRequired(TRUE);
 				$file->enableFileNameSelection("filename");
 				//$file->setSuffixes(array("doc","xls","png","jpg","gif","pdf"));
-				if ($_FILES["file"]["tmp_name"])
+				if( $_FILES["file"]["tmp_name"] && $file->checkInput() )
 				{
 					if (!file_exists($this->object->getSuggestedSolutionPath())) ilUtil::makeDirParents($this->object->getSuggestedSolutionPath());
+					
 					$res = ilUtil::moveUploadedFile($_FILES["file"]["tmp_name"], $_FILES["file"]["name"], $this->object->getSuggestedSolutionPath() . $_FILES["file"]["name"]);
 					if ($res)
 					{
+						ilUtil::renameExecutables($this->object->getSuggestedSolutionPath());
+						
 						// remove an old file download
 						if (is_array($solution_array["value"])) @unlink($this->object->getSuggestedSolutionPath() . $solution_array["value"]["name"]);
 						$file->setValue($_FILES["file"]["name"]);
@@ -1246,6 +1270,7 @@ abstract class assQuestionGUI
 					}
 					else
 					{
+						// BH: $res as info string? wtf? it holds a bool or something else!!?
 						ilUtil::sendInfo($res);
 					}
 				}
@@ -1748,5 +1773,10 @@ abstract class assQuestionGUI
 		);
 	}
 
+	public function showHints()
+	{
+		global $ilCtrl;
 
+		$ilCtrl->redirectByClass('ilAssQuestionHintsGUI', ilAssQuestionHintsGUI::CMD_SHOW_LIST);
+	}	
 }
